+++ bank/src/main/java/com/banksalad/collectmydata/bank/publishment/deposit/dto/DepositAccountTransactionsProtoResponse.java
@@ -3,7 +3,7 @@ package com.banksalad.collectmydata.bank.publishment.deposit.dto;
 import com.banksalad.collectmydata.common.util.DateUtil;
 import com.github.banksalad.idl.apis.v1.collectmydata.CollectmydatabankProto.BankDepositAccountTransaction;
 import com.github.banksalad.idl.apis.v1.collectmydata.CollectmydatabankProto.ListBankDepositAccountTransactionsResponse;
-import com.google.protobuf.Int64Value;
+import com.google.protobuf.Int32Value;
 import com.google.protobuf.StringValue;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -34,7 +34,7 @@ public class DepositAccountTransactionsProtoResponse {
             .setTransClass(depositAccountTransactionResponse.getTransClass())
             .setTransAmt3F(depositAccountTransactionResponse.getTransAmt().longValue())
             .setBalanceAmt3F(depositAccountTransactionResponse.getBalanceAmt().longValue())
-            .setPaidInCnt(Int64Value.newBuilder().setValue(depositAccountTransactionResponse.getPaidInCnt()).build())
+            .setPaidInCnt(Int32Value.newBuilder().setValue(depositAccountTransactionResponse.getPaidInCnt()).build())
             .setCreatedAtMs(DateUtil.utcLocalDateTimeToEpochMilliSecond(depositAccountTransactionResponse.getCreatedAt()))
             .setUpdatedAtMs(DateUtil.utcLocalDateTimeToEpochMilliSecond(depositAccountTransactionResponse.getUpdatedAt()))
             .build())
